requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pymongo==4.6.0
requests==2.31.0
rapidfuzz==3.5.2
kronoslabs
numpy<2.0.0
python-multipart
